初始配置时增加GUI，更换UI显示的依赖库

+++ auto_sign.py
@@ -3,7 +3,8 @@ from pyvirtualdisplay import Display
 import time
 from interval import Interval
 import requests
-import win32api, win32con
+import tkinter as tk
+import tkinter.messagebox as msgbox
 import sys
 import os.path
 
@@ -14,6 +15,12 @@ headers = {
     "Accept-Encoding": "gzip, deflate, br",
     "Connection": "keep-alive"
 }
+
+def default_config():
+    new_config = open("./config_default.txt", 'w', encoding="utf-8")
+    new_config.write(
+        'username:"XXXXXXXX,XXXXXXXX"\npw:"12345678,12345678"\nmailto:"XXXXXXXXX@qq.com"\nhidden:1\nemail:1')
+    new_config.close()
 
 def write(data):
     File = open('log.txt', mode='a', encoding="utf-8")
@@ -60,7 +67,8 @@ def daka(user, pw, mailto, hidden):
         except:
             write("在登陆页面出错，可能是账号密码错误")
             print("在登陆页面出错，可能是账号密码错误")
-            win32api.MessageBox(0, user + "登陆失败，可能是账号或密码错误，请检查您的帐号与密码是否正确且一一对应", "登陆失败", win32con.MB_ICONWARNING)
+            msgbox.showerror("登陆失败", user + "登陆失败，可能是账号或密码错误，请检查您的帐号与密码是否正确且一一对应")
+            #  win32api.MessageBox(0, user + "登陆失败，可能是账号或密码错误，请检查您的帐号与密码是否正确且一一对应", "登陆失败", win32con.MB_ICONWARNING)
             driver.quit()
             if hidden == 1:
                 window.stop()
@@ -75,11 +83,13 @@ def daka(user, pw, mailto, hidden):
                 write(user + "打卡成功")
                 print(user + "打卡成功，如果是首次使用，建议人工查看是否已打卡")
             except:
-                win32api.MessageBox(0, user + "打卡失败，请手动打卡", "打卡失败", win32con.MB_ICONWARNING)
+                msgbox.showerror("打卡失败", user + "打卡失败，请手动打卡")
+                #  win32api.MessageBox(0, user + "打卡失败，请手动打卡", "打卡失败", win32con.MB_ICONWARNING)
                 write(user + "打卡失败")
                 print(user + "打卡失败，请手动打卡")
         else:
-            win32api.MessageBox(0, user + "登录失败，可能是浏览器未正确跳转，请重试或在GitHub反馈", "登陆失败", win32con.MB_ICONWARNING)
+            msgbox.showerror("登录失败", user + "登录失败，可能是浏览器未正确跳转，请重试或在GitHub反馈")
+            #  win32api.MessageBox(0, user + "登录失败，可能是浏览器未正确跳转，请重试或在GitHub反馈", "登陆失败", win32con.MB_ICONWARNING)
             write("登陆失败，可能是浏览器未正确跳转")
             print("登陆失败，可能是浏览器未正确跳转")
 
@@ -93,7 +103,7 @@ def daka(user, pw, mailto, hidden):
     write("浏览器已关闭")
     print("浏览器已关闭")
 
-    if mailto != 0:
+    if mailto != 0:  #  递送打卡报告
         with open('./log.txt', encoding='utf-8') as file:
             content = file.read()
         content = content.split("#")[-1]
@@ -103,6 +113,19 @@ def daka(user, pw, mailto, hidden):
         print(r.text)
 
 
+
+#  判断是否展示条约
+if not os.path.isfile("./readed"):
+    #  rule = win32api.MessageBox(0, "使用本程序带来的后果，YujioNako与Pro-Ivan概不负责，如不同意该条件请勿使用本程序！", "声明", win32con.MB_YESNO)
+    rule = msgbox.askyesno("声明", "使用本程序带来的后果，YujioNako与Pro-Ivan概不负责，如不同意该条件请勿使用本程序！\n是否同意该规则？")
+    if rule == 0:
+        sys.exit()
+    elif rule == 1:
+        file = open("./readed", 'w', encoding="utf-8")
+        file.write("本文档的存在意味着您接受我们的条约，即使用本程序带来的后果，YujioNako与Pro-Ivan概不负责。敬请知悉。")
+        file.close()
+
+
 with open("./geckodriver.log", 'w') as file:
     file.truncate(0)
 with open("./log.txt", 'w') as file:
@@ -110,14 +133,40 @@ with open("./log.txt", 'w') as file:
 print("log初始化成功")
 write("读取config")
 print("正在读取config.txt")
-if not os.path.isfile("./config.txt"):
-    print("初次使用吗？是的话请告诉我们您的BUPT账号和密码吧！")
-    user = input("账号：")
-    pw = input("密码：")
-    new_config = open("./config.txt", 'w', encoding="utf-8")
-    new_config.write('username:"' + user + '"\npw:"' + pw + '"\nmailto:""\nhidden:1\nemail:0')
-    new_config.close()
-    print("基本设置录入成功！更多设置请自行查看config.txt")
+
+if not os.path.isfile("./config.txt"):  #  config初始设置GUI
+    #  print("初次使用吗？是的话请告诉我们您的BUPT账号和密码吧！")
+    first_time = msgbox.askyesno("欢迎使用", "欢迎使用由YujioNako编写的BUPT自动打卡程序！您似乎是第一次使用本程序，需要钰中帮您配置账密吗？或者您也可以选“否”，然后根据我们提供的config_default.txt自行配置后将文件名改为config.txt")
+    if first_time == 0:
+        default_config()
+        sys.exit()
+    #  user = input("账号：")
+    #  pw = input("密码：")
+    root = tk.Tk()
+    root.geometry('300x240')
+
+    def input2config():
+        user = b0.get()
+        pw = b1.get()
+        new_config = open("./config.txt", 'w', encoding="utf-8")
+        new_config.write('username:"' + user + '"\npw:"' + pw + '"\nmailto:""\nhidden:1\nemail:0')
+        new_config.close()
+        msgbox.showinfo("config已设置", "基本设置录入成功！更多设置请自行查看config.txt")
+        print("基本设置录入成功！更多设置请自行查看config.txt")
+        root.destroy()
+
+    b3 = tk.Label(root, text="学号：")
+    b3.pack()
+    b0 = tk.Entry(root, text='学号', width=20)
+    b0.pack()
+    b4 = tk.Label(root, text="密码：")
+    b4.pack()
+    b1 = tk.Entry(root, text='密码', width=20)
+    b1.pack()
+    b2 = tk.Button(root, text='确 认', command=input2config)
+    b2.pack()
+    root.mainloop()
+
 try:
     with open('./config.txt', 'r') as file:
         l = file.readlines()  # 按行读取TXT文件，都是字符串类型
@@ -130,34 +179,22 @@ try:
             else:
                 s0[1] = int(s0[1])  # 说明原本值是整型，强制类型转换
             d[s0[0]] = s0[1]  # 键值对添加到字典中
+    user = d['username']
+    user = user.split(",")
+    pw = d['pw']
+    pw = pw.split(",")
+    mailto = d['mailto']
+    mailto = mailto.split(",")
+    hidden = d['hidden']
+    email = d['email']
+    print("完成config的读取")
 except:
-    win32api.MessageBox(0, "可能是config.txt内格式错误，请从目录内找到config_default.txt进行参考", "config异常", win32con.MB_ICONWARNING)
-    new_config = open("./config_default.txt", 'w', encoding="utf-8")
-    new_config.write('username:"XXXXXXXX,XXXXXXXX"\npw:"12345678,12345678"\nmailto:"XXXXXXXXX@qq.com"\nhidden:1\nemail:1')
-    new_config.close()
+    msgbox.showerror("config异常", "可能是config.txt内格式错误，请从目录内找到config_default.txt进行参考")
+    #  win32api.MessageBox(0, "可能是config.txt内格式错误，请从目录内找到config_default.txt进行参考", "config异常", win32con.MB_ICONWARNING)
+    default_config()
     write("config异常")
     print("config异常")
     sys.exit()
-
-user = d['username']
-user = user.split(",")
-pw = d['pw']
-pw = pw.split(",")
-mailto = d['mailto']
-mailto = mailto.split(",")
-hidden = d['hidden']
-email = d['email']
-print("完成config的读取")
-
-#  判断是否展示条约
-if not os.path.isfile("./readed"):
-    rule = win32api.MessageBox(0, "使用本程序带来的后果，YujioNako与Pro-Ivan概不负责，如不同意该条件请勿使用本程序！", "声明", win32con.MB_YESNO)
-    if rule == 7:
-        sys.exit()
-    elif rule == 6:
-        file = open("./readed", 'w', encoding="utf-8")
-        file.write("本文档的存在意味着您接受我们的条约，即使用本程序带来的后果，YujioNako与Pro-Ivan概不负责。敬请知悉。")
-        file.close()
 
 
 while True:
@@ -176,7 +213,8 @@ while True:
 
         def indexing(lst, index):
             if lst[index:] == []:
-                win32api.MessageBox(0, "list index out of range", "IndexError", win32con.MB_ICONWARNING)
+                msgbox.showwarning("IndexError", "list index out of range")
+                #  win32api.MessageBox(0, "list index out of range", "IndexError", win32con.MB_ICONWARNING)
                 print("IndexError: list index out of range")
 
         for i in range(len(user)):
@@ -188,7 +226,8 @@ while True:
                 if email == 0:
                     daka(user[i], pw[i], 0, hidden)
             except:
-                win32api.MessageBox(0, "可能是因为config.txt内密码/邮箱数组越界，请检查config.txt", "错误", win32con.MB_ICONWARNING)
+                msgbox.showerror("错误", "可能是因为config.txt内密码/邮箱数组越界，请检查config.txt")
+                #  win32api.MessageBox(0, "可能是因为config.txt内密码/邮箱数组越界，请检查config.txt", "错误", win32con.MB_ICONWARNING)
                 print("出现错误，可能是因为密码/邮箱数组越界")
                 write("出现错误，可能是因为密码/邮箱数组越界")
                 sys.exit()
